requirements.txt
Django==1.11.4
djangorestframework==3.6.4
gunicorn==19.7.1
pytz==2017.2
